+++ app/reports/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class DenunciasConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'app.denuncias'
+    name = 'app.reports'
