+++ app/reports/models.py
@@ -1,0 +1,12 @@
+# reports/models.py
+from django.db import models
+
+class Report(models.Model):
+    user = models.ForeignKey('users.UserBase', on_delete=models.CASCADE)
+    comment = models.ForeignKey('schools.Comment', on_delete=models.CASCADE)
+    reason = models.TextField(null=False, blank=False)
+    is_reviewed = models.BooleanField(default=False)
+    created_at = models.DateTimeField(auto_now_add=True)
+
+    def __str__(self):
+        return f'Report by {self.user} on comment {self.comment.id}'
